requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
# Config-driven dataset path resolution
python-dotenv>=1.0.0
openai>=1.0.0
python-multipart>=0.0.6
requests>=2.28.0
tiktoken>=0.5.0
pytest>=7.0.0
pytest-cov>=4.1.0
httpx>=0.24.0
httpx_sse>=0.4.0
sse-starlette>=2.1.3
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
networkx>=3.0
scipy>=1.11.0